+++ app/src/main/java/org/eyeseetea/malariacare/layout/adapters/survey/DynamicTabAdapter.java
@@ -163,10 +163,10 @@ public class DynamicTabAdapter extends BaseAdapter implements ITabAdapter {
                 Option selectedOption=(Option)view.getTag();
                 Question question=navigationController.getCurrentQuestion();
                 //Uncomment to read in the console the workflow values on saveValue
-                //Note: it is commented becouse showValues needs a sql query and make the application more slowy.
-                //Log.d(TAG, Session.getSurvey().showValues());
+                //Note: it is commented becouse printValues needs a sql query and make the application more slowy.
+                //Log.d(TAG, Session.getSurvey().printValues());
                 ReadWriteDB.saveValuesDDL(question, selectedOption);
-                //Log.d(TAG, Session.getSurvey().showValues());
+                //Log.d(TAG, Session.getSurvey().printValues());
                 ViewGroup vgTable = (ViewGroup) view.getParent().getParent();
                 for (int rowPos = 0; rowPos < vgTable.getChildCount(); rowPos++) {
                     ViewGroup vgRow = (ViewGroup) vgTable.getChildAt(rowPos);
